requirements.txt
alembic==1.8.0
aniso8601==9.0.1
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
Flask==2.1.2
Flask-API==3.0.post1
Flask-API.yandex==0.6.2.1
Flask-JWT-Extended==4.4.2
Flask-Login==0.6.1
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
psycopg2-binary==2.9.3
PyJWT==2.4.0
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.39
stripe==3.5.0
urllib3==1.26.10
Werkzeug==2.1.2
zipp==3.8.0
pip~=22.1.1
wheel~=0.37.1
setuptools~=62.3.2
cloudinary~=1.29.0